+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0]
+
 ### Added
 
+- Custom validators for configuration elements that run after all the standard
+  validations for that element are done.
 - Automatically ignore and prune any key on an ObjectElement that starts with
   the **$** sign, as these normally contain meta-data like comments or schema
   links.
@@ -33,7 +37,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial version of configuration parser.
 
-[Unreleased]: https://github.com/mischareitsma/config-parser-ts/compare/v0.3.0...HEAD
-[0.2.0]: https://github.com/mischareitsma/config-parser-ts/compare/v0.2.0...v0.3.0
+[Unreleased]: https://github.com/mischareitsma/config-parser-ts/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/mischareitsma/config-parser-ts/compare/v0.3.0...v0.4.0
+[0.3.0]: https://github.com/mischareitsma/config-parser-ts/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/mischareitsma/config-parser-ts/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/mischareitsma/config-parser-ts/releases/tag/v0.1.0